fix bugs

+++ src/TuioNet/Tuio20/Tuio20Symbol.cs
@@ -54,9 +54,9 @@ namespace TuioNet.Tuio20
             get
             {
                 var message = new OSCMessage("/tuio2/sym");
-                message.Append(SessionId);
-                message.Append(TypeUserId);
-                message.Append(ComponentId);
+                message.Append((int)SessionId);
+                message.Append((int)TypeUserId);
+                message.Append((int)ComponentId);
                 message.Append(Group);
                 message.Append(Data);
                 return message;
